[Silver I] Title: 곱셈, Time: 80 ms, Memory: 11664 KB -BaekjoonHub

+++ 백준/Silver/1629. 곱셈/곱셈.java
@@ -34,6 +34,6 @@ public class Main {
 				sum = sum.multiply(BigInteger.valueOf(arr[N-i])).remainder(BigInteger.valueOf(C));
 			}
 		}
-		System.out.println(sum);
+		System.out.println(sum.longValue());
 	}
 }
